--- src/App.jsx
@@ -1,13 +1,0 @@
-import "./App.css";
-
-import WebBase from "./components/WebBase.jsx";
-
-function App() {
-  return (
-    <>
-      <WebBase />
-    </>
-  );
-}
-
-export default App;
