--- src/components/NoteContainer.jsx
@@ -1,79 +1,0 @@
-function NoteContainer() {
-  return (
-    <>
-      <div className="flex-1 overflow-y-scroll p-10 sm:grid sm:grid-cols-3 sm:grid-rows-10 sm:gap-4">
-        <div
-          contentEditable
-          className="col-span-1 row-span-10 overflow-y-scroll rounded-md border border-dashed bg-[#F5F7F8] p-2 text-slate-900"
-        >
-          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Vitae
-          reprehenderit rerum deleniti modi voluptates ipsam delectus aliquam
-          fugiat, voluptatibus odit sequi doloribus perspiciatis. Eos id error
-          facilis soluta impedit voluptatibus. Saepe cumque quis, cupiditate
-          numquam dolorum distinctio. Accusamus, expedita laudantium iusto dicta
-          provident optio ipsa eius, similique voluptatem, animi soluta omnis!
-          Perferendis sapiente consectetur hic laborum impedit, ipsa quibusdam
-          amet. Harum dolorum commodi esse consequatur beatae! Aliquid
-          consequuntur deserunt facere molestias ratione amet neque dolores
-          quibusdam quae incidunt, excepturi minus hic aliquam, vitae nobis
-          accusamus. Tempora labore voluptatem molestiae eaque? Quidem minima,
-          ut modi earum accusamus illo velit illum eos rerum inventore eius quis
-          ex perferendis molestias numquam. Unde dignissimos consectetur dolores
-          impedit eligendi harum iusto, architecto eaque voluptate cumque?
-          Repellat vero expedita quisquam commodi voluptatem veritatis natus
-          beatae pariatur ex rerum, atque suscipit consectetur magni quibusdam.
-          Similique repellendus quae at incidunt nam porro doloremque
-          recusandae, odio veniam quaerat! Pariatur.
-        </div>
-        <div
-          contentEditable
-          className="col-span-2 row-span-7 overflow-y-scroll rounded-md border border-dashed bg-[#F5F7F8] p-2 text-slate-900"
-        >
-          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Vitae
-          reprehenderit rerum deleniti modi voluptates ipsam delectus aliquam
-          fugiat, voluptatibus odit sequi doloribus perspiciatis. Eos id error
-          facilis soluta impedit voluptatibus. Saepe cumque quis, cupiditate
-          numquam dolorum distinctio. Accusamus, expedita laudantium iusto dicta
-          provident optio ipsa eius, similique voluptatem, animi soluta omnis!
-          Perferendis sapiente consectetur hic laborum impedit, ipsa quibusdam
-          amet. Harum dolorum commodi esse consequatur beatae! Aliquid
-          consequuntur deserunt facere molestias ratione amet neque dolores
-          quibusdam quae incidunt, excepturi minus hic aliquam, vitae nobis
-          accusamus. Tempora labore voluptatem molestiae eaque? Quidem minima,
-          ut modi earum accusamus illo velit illum eos rerum inventore eius quis
-          ex perferendis molestias numquam. Unde dignissimos consectetur dolores
-          impedit eligendi harum iusto, architecto eaque voluptate cumque?
-          Repellat vero expedita quisquam commodi voluptatem veritatis natus
-          beatae pariatur ex rerum, atque suscipit consectetur magni quibusdam.
-          Similique repellendus quae at incidunt nam porro doloremque
-          recusandae, odio veniam quaerat! Pariatur.
-        </div>
-        <div
-          contentEditable
-          className="col-span-2 col-start-2 row-span-3 row-start-8 overflow-y-scroll rounded-md border  border-dashed bg-[#F5F7F8] p-2 text-slate-900"
-        >
-          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Vitae
-          reprehenderit rerum deleniti modi voluptates ipsam delectus aliquam
-          fugiat, voluptatibus odit sequi doloribus perspiciatis. Eos id error
-          facilis soluta impedit voluptatibus. Saepe cumque quis, cupiditate
-          numquam dolorum distinctio. Accusamus, expedita laudantium iusto dicta
-          provident optio ipsa eius, similique voluptatem, animi soluta omnis!
-          Perferendis sapiente consectetur hic laborum impedit, ipsa quibusdam
-          amet. Harum dolorum commodi esse consequatur beatae! Aliquid
-          consequuntur deserunt facere molestias ratione amet neque dolores
-          quibusdam quae incidunt, excepturi minus hic aliquam, vitae nobis
-          accusamus. Tempora labore voluptatem molestiae eaque? Quidem minima,
-          ut modi earum accusamus illo velit illum eos rerum inventore eius quis
-          ex perferendis molestias numquam. Unde dignissimos consectetur dolores
-          impedit eligendi harum iusto, architecto eaque voluptate cumque?
-          Repellat vero expedita quisquam commodi voluptatem veritatis natus
-          beatae pariatur ex rerum, atque suscipit consectetur magni quibusdam.
-          Similique repellendus quae at incidunt nam porro doloremque
-          recusandae, odio veniam quaerat! Pariatur.
-        </div>
-      </div>
-    </>
-  );
-}
-
-export default NoteContainer;
