--- tailwind.config.js
@@ -1,11 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
-  theme: {
-    extend: {},
-    fontFamily:{
-      main:["PT Sans","sans-serif"]
-    }
-  },
-  plugins: [],
-};
